+++ ArrayTraits/src/ArrayAccess.php
@@ -32,7 +32,7 @@ trait ArrayAccess
     #[\ReturnTypeWillChange]
     public function offsetGet($offset)
     {
-        return isset($this->items[$offset]) ? $this->items[$offset] : null;
+        return $this->items[$offset] ?? null;
     }
 
     /**