+++ CHANGELOG.md
@@ -10,6 +10,7 @@
    * Updated code for PHP 7.2
    * Added support for PHP 8.4 to handle deprecations
    * Added support for `ordering@` on fields with `import@`
+   * Fixed deprecated messages
 
 # v1.6.5
 ## 05/09/2023